pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "weaklink-platform"
version = "0.1.0"
description = "Internal platform automation for WeakLink Labs"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.11,<1",
]

[tool.setuptools.packages.find]
include = ["weaklink_platform*"]

[tool.pytest.ini_options]
testpaths = ["tests/platform"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "I001"]
